assignment-01/requirements.txt
numpy==1.14.2
Pillow==5.0.0
pkg-resources==0.0.0
scipy==1.0.1
